Fix the doc link
Changed the link to the docs so that it would not give user 404. It was https://vaadin.com/docs/business-app/overview.html which does not exist, currently correct one is https://vaadin.com/docs/v13/business-app/overview.html. Perhaps even better would be something like https://vaadin.com/docs/latest/business-app/overview.html so that we do not have to remember to update the doc link when new platform version is out