Fix AccountsIT test
Fixes https://github.com/vaadin/business-app-starter-flow/issues/87
HashMaps used for storing data are populated at start-up time. Therefore, it's not certainly known what would be the values of each row.
This pull request make test independable of user data and verifies instead that the same value for availability is displayed in both Grid and GridDetails view for a company.